+++ src/Subscriber/OrderSubscriber.php
@@ -6,6 +6,7 @@ use Shopware\Core\Checkout\Cart\Event\CheckoutOrderPlacedEvent;
 use Shopware\Core\Checkout\Order\OrderEntity;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
+use Tinect\Matomo\Service\StaticHelper;
 use Tinect\Matomo\Tracking\ServerSideTracker;
 use Tinect\Matomo\Tracking\TrackingPayloadBuilder;
 use Tinect\Matomo\Tracking\VisitorIdResolver;
@@ -93,6 +94,6 @@ class OrderSubscriber implements EventSubscriberInterface
             return '';
         }
 
-        return $request->getSchemeAndHttpHost() . $request->getRequestUri();
+        return StaticHelper::buildStorefrontUrl($request);
     }
 }